C++ code matches C code.
C++ code wasn't matching C code when the Autolev formulation changed to only
have one rigid body of mass m, instead of two rigid bodies of mass ma and mb.
Not sure why that approach didn't work, but I have now ruled out the
possibility that there was an error in the C++ code instead of an error in the
autolev code.