Added output of omega in 'natural' frame coordinates.
This revealed that the angular velocity of the twin discs only has a component
about the contact line.  The next step is to reformulate the angular velocity
expression for the disc using only one generalized speed.  This should
eliminate the need for any nonholonomic constraint equations, and hopefully
should simplify the equations.