Converted code to use only one generalized speed.
Everything seems to be identical, but comparison is a bit difficult since now
it isn't trivial to specify the same initial bodyfixed angular velocity;
instead the angular velocity about the contact line must be specified and they
are a bit different.  Otherwise, a reduction in the complexity of the equations
was noticed in the form of about 20 fewer Z variables needed.